refactor: move strip_index_notation to mappings/utils.py
Also run the doctest test cases.
